Renamed CompGraphNode to Tensor and added .realize()
